refactor color related functions out of index
